backend/requirements.txt
# TKA Voice Agent Backend Dependencies

# FastAPI and Web Framework
fastapi==0.115.12
uvicorn[standard]==0.28.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# AI Services (Compatible versions)
# openai==1.54.4
google-generativeai==0.8.3

# Telephony
twilio==9.3.7

# Audio Processing & Transcription
# openai-whisper==20231117
# torch==2.1.0
# torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Task Queue and Background Jobs
celery==5.4.0
redis==5.2.0

# Email
fastapi-mail==1.4.1

# Monitoring and Logging
python-json-logger==2.0.7
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Environment and Configuration
python-dotenv==1.0.1

# Date and Time
python-dateutil==2.9.0.post0
pytz==2024.2

# Validation and Utilities
validators==0.34.0
phonenumbers==8.13.48

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0 